backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
email-validator==2.1.0
requests==2.31.0
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
argon2-cffi>=23.1.0
python-jose[cryptography]==3.3.0
redis>=5.0.0
